Support using a custom Blockstore
Usually (in Kubo), the Blockstore is just built on top of the datastore.

There are, however, blockstore implementations (i.e StoreTheHash) that are
optimized blockstores and not general purpose datastores.

This commit allows providing a custom blockstore, but initializes everything
as before when set to nil. The signature of New() changes accordingly. A new
option UncachedBlockstore, allows disabling the CachedBlockstore-wrapping that
also happens by default, since some datastores/blockstores include their own
caching and bloom-filters, so this can potentially be making things slower.